feat(mobile): combine friends pages
