+++ src/components/main/files/styles.scss
@@ -11,12 +11,14 @@
     display: inline-flex;
     flex-direction: column;
   }
+
   .flex-row {
     display: flex;
     flex-direction: row;
     flex-wrap: wrap;
     width: 100%;
   }
+
   .mobile-usage {
     justify-self: flex-start;
     display: flex;
@@ -30,6 +32,7 @@
     margin-bottom: 0.2rem;
     border-bottom: 1px solid var(--theme-borders);
   }
+
   .mobile-nav {
     margin: 1rem;
   }
