+++ src/components/main/friends/request/style.scss
@@ -15,6 +15,8 @@
   }
 
   .who {
+    display: flex;
+    align-items: center;
     height: 40px;
     text-align: left;
     flex-grow: 1;
@@ -32,6 +34,6 @@
     display: flex;
     flex-shrink: 0;
     height: 40px;
-    gap: 16px;
+    gap: 0.5rem;
   }
 }
