+++ src/components/main/friends/sidebar/styles.scss
@@ -14,33 +14,5 @@
       padding-left: 0;
       display: inline-flex;
     }
-
-    .code {
-      width: 100%;
-      margin-bottom: 1rem;
-      .button {
-        min-width: 100%;
-      }
-    }
-
-    .add {
-      width: 100%;
-      display: inline-flex;
-      margin-bottom: 1rem;
-
-      .icon-input {
-        flex: 1;
-      }
-
-      .button {
-        margin-left: 8px;
-      }
-
-      @media only screen and (min-width: 600px) {
-        .show-friends-button {
-          display: none;
-        }
-      }
-    }
   }
 }
