+++ src/components/main/friends/styles.scss
@@ -16,6 +16,10 @@
     overflow-y: scroll;
     flex-direction: column;
 
+    .scroll-container {
+      overflow: hidden;
+    }
+
     .main {
       display: flex;
 
@@ -37,7 +41,21 @@
         display: flex;
         flex-direction: column;
         justify-content: space-between;
+        color: var(--theme-text-muted);
       }
+    }
+
+    .mobile-wrap {
+      border-bottom: 1px solid var(--theme-borders);
+      padding: 1rem;
+
+      @media (min-width: 601px) {
+        display: none;
+      }
+    }
+
+    .mobile-nav {
+      padding: 1rem;
     }
   }
