+++ src/components/reusable/page_header/styles.scss
@@ -2,7 +2,7 @@
   margin-bottom: 0.5rem;
   border-bottom: 1px solid var(--theme-borders);
 
-  @media only screen and (min-width: 600px) {
+  @media only screen and (min-width: 601px) {
     &.header-hidden {
       display: none;
     }
