+++ src/ui_kit/src/input/styles.scss
@@ -1,5 +1,7 @@
 .input-container {
   position: relative;
+  display: flex;
+  flex-grow: 1;
 
   .input {
     min-width: 100px;
